requirements.txt
streamlit
onnxruntime
numpy
Pillow